updates for the grey room
This update does a few things. First it integrates the offline routine with PsychoPy's parallel library so that we can record it on the ANT EEG system. Second, it fixes a few bugs around timings and visualization of the grid. The routine highlights the target in red now, so may not be appropriate for people with colorblindness.